% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezpickr-package.R
\docType{package}
\name{ezpickr-package}
\alias{ezpickr}
\alias{ezpickr-package}
\title{ezpickr: Import Various Data File Types as a Rectangular Form Using a File Picker Dialogue Box}
\description{
Easy data importing for the most frequently used file formats in a 'tibble' form.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/ezpickr}
  \item Report bugs at \url{https://github.com/jooyoungseo/ezpickr/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jooyoung@psu.edu}

Authors:
\itemize{
  \item Soyoung Choi \email{sxc940@psu.edu}
}

Other contributors:
\itemize{
  \item hyun seung Lee \email{go4arsenal@gmail.com} [contributor]
  \item Jonghoen Kim \email{cysics@daum.net} [contributor]
}

}
\keyword{internal}
