% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezpickr-package.R
\docType{package}
\name{ezpickr-package}
\alias{ezpickr}
\alias{ezpickr-package}
\title{ezpickr: Easy Data Import Using GUI File Picker and Seamless Communication Between an Excel and R}
\description{
Choosing any rectangular data file using interactive GUI dialog box, and seamlessly manipulating tidy data between an 'Excel' window and R session.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/ezpickr}
  \item Report bugs at \url{https://github.com/jooyoungseo/ezpickr/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jooyoung@psu.edu} (\href{https://orcid.org/0000-0002-4064-6012}{ORCID})

Authors:
\itemize{
  \item Soyoung Choi \email{sxc940@psu.edu} (\href{https://orcid.org/0000-0002-0998-3352}{ORCID})
}

Other contributors:
\itemize{
  \item hyun seung Lee \email{go4arsenal@gmail.com} [contributor]
  \item Jonghoen Kim \email{cysics@daum.net} [contributor]
}

}
\keyword{internal}
