% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Convenient function for viewing any rectangular data file in an Excel window.}
\usage{
view(x, ...)
}
\arguments{
\item{x}{An object of class data.frame, matrix, table or vector.}

\item{...}{Any additional arguments available for \link[writexl]{write_xlsx}.}
}
\value{
Data object opened in a preferable spreadsheet application window.
}
\description{
You can use this function for loading any data.frame, data_frame, tbl_df, matrix, table, vector objects into your system-default spreadsheet software (e.g., Excel). This function has been inspired by \code{\link[BrailleR]{DataViewer}} and has implemented \code{\link[writexl]{write_xlsx}} instead of the default \code{\link[utils]{write.csv}} for a better performance.
}
\details{
view


See example below.
}
\examples{
if(interactive()) {
library(ezpickr)
data(airquality)
str(airquality)
## View your data object in your spreadsheet software:
view(airquality)
# Then, when necessary, you can modify the opened data in the spreadsheet and save it as a new data.

# You can pass a list object to the `view()` function like below:
l <- list(iris = iris, mtcars = mtcars, chickwts = chickwts, quakes = quakes)
view(l)
# Then, each list item will appear in your Excel window sheet by sheet.
}
}
\author{
JooYoung Seo, \email{jooyoung@psu.edu}

Soyoung Choi, \email{sxc940@psu.edu}
}
\keyword{view}
